refactor schemas to share asset enums
