+++ src/market-data/index.ts
@@ -19,3 +19,4 @@ export * as movers from './movers'
 export * as options from './options'
 export * as priceHistory from './priceHistory'
 export * as quotes from './quotes'
+export * as shared from './shared'
