+++ src/market-data/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './instrument-asset-type.schema'
