+++ src/schemas/index.ts
@@ -4,7 +4,9 @@ export * from '../market-data/movers/schema'
 export * from '../market-data/options/schema'
 export * from '../market-data/priceHistory/schema'
 export * from '../market-data/quotes/schema'
+export * from '../market-data/shared'
 export * from '../trader/accounts/schema'
 export * from '../trader/orders/schema'
 export * from '../trader/transactions/schema'
 export * from '../trader/user-preference/schema'
+export * from '../trader/shared'
