+++ src/trader/accounts/schema.ts
@@ -1,15 +1,5 @@
 import { z } from 'zod'
-import { assetType } from '../transactions/schema'
-
-const ApiCurrencyType = z.enum(['USD', 'CAD', 'EUR', 'JPY'])
-
-const AccountAPIOptionDeliverable = z.object({
-	symbol: z.string(),
-	deliverableUnits: z.number(),
-	apiCurrencyType: ApiCurrencyType,
-	assetType: assetType,
-})
-type AccountAPIOptionDeliverable = z.infer<typeof AccountAPIOptionDeliverable>
+import { assetType, ApiCurrencyType, AccountAPIOptionDeliverable } from '../shared'
 
 export const AccountsBaseInstrument = z.object({
 	assetType: assetType,
