+++ src/trader/index.ts
@@ -17,3 +17,4 @@ export * as accounts from './accounts'
 export * as orders from './orders'
 export * as transactions from './transactions'
 export * as userPreference from './user-preference'
+export * as shared from './shared'
