+++ src/trader/shared/account-api-option-deliverable.schema.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod'
+import { ApiCurrencyType } from './api-currency-type.schema'
+import { assetType } from './asset-type.schema'
+
+export const AccountAPIOptionDeliverable = z.object({
+        symbol: z.string(),
+        deliverableUnits: z.number(),
+        apiCurrencyType: ApiCurrencyType,
+        assetType: assetType,
+})
+export type AccountAPIOptionDeliverable = z.infer<typeof AccountAPIOptionDeliverable>
