+++ src/trader/shared/api-currency-type.schema.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod'
+
+export const ApiCurrencyType = z.enum(['USD', 'CAD', 'EUR', 'JPY'])
+export type ApiCurrencyType = z.infer<typeof ApiCurrencyType>
