+++ src/trader/transactions/schema.ts
@@ -1,24 +1,10 @@
 import { z } from 'zod'
 import {
-	isoDateTimeSchema,
-	DateFormatType,
-	dateTransformer,
+        isoDateTimeSchema,
+        DateFormatType,
+        dateTransformer,
 } from '../../utils/date-utils'
-
-export const assetType = z.enum([
-	'EQUITY',
-	'MUTUAL_FUND',
-	'OPTION',
-	'FUTURE',
-	'FOREX',
-	'INDEX',
-	'CASH_EQUIVALENT',
-	'FIXED_INCOME',
-	'PRODUCT',
-	'CURRENCY',
-	'COLLECTIVE_INVESTMENT',
-])
-export type AssetType = z.infer<typeof assetType>
+import { assetType } from '../shared'
 
 export const TransactionType = z.enum([
 	'TRADE',
